Minor refactor Firebase Service classes to infrastructure package.
