+++ SFFoodTruck/app/src/main/java/com/codepath/com/sffoodtruck/ui/notification/FoodTruckNotification.java
@@ -1,30 +1,21 @@
-package com.codepath.com.sffoodtruck.data.remote.service;
+package com.codepath.com.sffoodtruck.ui.notification;
 
 import android.app.NotificationManager;
 import android.content.Context;
 import android.support.v4.app.NotificationCompat;
 
 import com.codepath.com.sffoodtruck.R;
-import com.google.firebase.messaging.FirebaseMessagingService;
 import com.google.firebase.messaging.RemoteMessage;
 
-import java.util.Map;
+/**
+ * Created by robl2e on 10/20/17.
+ */
 
-public class FCMMessageHandlerService extends FirebaseMessagingService {
-    public static final int MESSAGE_NOTIFICATION_ID = 435345;
-
-    @Override
-    public void onMessageReceived(RemoteMessage remoteMessage) {
-        Map<String, String> data = remoteMessage.getData();
-        String from = remoteMessage.getFrom();
-
-        RemoteMessage.Notification notification = remoteMessage.getNotification();
-        createNotification(notification);
-    }
+public class FoodTruckNotification {
+    private static final int MESSAGE_NOTIFICATION_ID = 435345;
 
     // Creates notification based on title and body received
-    private void createNotification(RemoteMessage.Notification notification) {
-        Context context = getBaseContext();
+    public void createNotification(Context context, RemoteMessage.Notification notification) {
         NotificationCompat.Builder mBuilder = new NotificationCompat.Builder(context)
                 .setSmallIcon(R.mipmap.ic_launcher).setContentTitle(notification.getTitle())
                 .setContentText(notification.getBody());
@@ -32,5 +23,4 @@ public class FCMMessageHandlerService extends FirebaseMessagingService {
                 .getSystemService(Context.NOTIFICATION_SERVICE);
         mNotificationManager.notify(MESSAGE_NOTIFICATION_ID, mBuilder.build());
     }
-
 }
